+++ packages/shared/hooks/useShortcut/index.ts
@@ -1,4 +1,4 @@
-import { onBeforeUnmount, ref, ShallowReactive, shallowReactive, watch, WatchHandle } from "vue"
+import { onBeforeUnmount, readonly, ref, ShallowReactive, shallowReactive, watch, WatchHandle } from "vue"
 import hotkeys from "hotkeys-js"
 import PQueue from "p-queue"
 import { uniqueId } from "../../misc"
@@ -32,18 +32,6 @@ export function useShortcut() {
     return true
   }
   hotkeys.setScope(scope)
-  function cleanAll() {
-    Array.from(handler.map(h => h.id)).forEach(clean)
-    handler.length = 0
-  }
-  function clean(id: string) {
-    const index = handler.findIndex(h => h.id === id)
-    if (index === -1) return
-    hotkeys.unbind(handler[index].key, scope)
-    handler[index].handler.stop()
-    handler[index].queue.clear()
-    handler.splice(index, 1)
-  }
   function listen(
     shortcut: string,
     callback: (active: boolean, key: string, ...args: unknown[]) => Promise<void> | void,
@@ -52,7 +40,29 @@ export function useShortcut() {
     const id = uniqueId()
     const key = ref(shortcut)
     const queue = new PQueue({ concurrency: 1 })
-    const listen = (val: string, old?: string) => {
+    const taskCount = ref(0)
+
+    const taskPending = ref(false)
+
+    queue.addListener("idle", () => {
+      taskPending.value = false
+      taskCount.value = 0
+    })
+    // Emitted every time the add method is called and the number of pending or queued tasks is increased.
+    queue.addListener("add", () => {
+      taskPending.value = true
+      taskCount.value++
+    })
+    // Emitted as each item is processed in the queue for the purpose of tracking progress.
+    queue.addListener("active", () => {
+      taskPending.value = true
+    })
+    // Emitted every time a task is completed and the number of pending or queued tasks is decreased.
+    // This is emitted regardless of whether the task completed normally or with an error.
+    queue.addListener("next", () => {
+      taskCount.value = Math.max(0, taskCount.value - 1)
+    })
+    const keyListener = (val: string, old?: string) => {
       if (!val) return
       if (val === old) return
       const current = handler.find(h => h.id === id)
@@ -67,14 +77,14 @@ export function useShortcut() {
         queue.add(async () => callback(event.type === "keydown", key))
       })
     }
-    const watcher = watch(key, listen)
+    const watcher = watch(key, keyListener)
     handler.push({
       id,
       key: key.value,
       handler: watcher,
       queue,
     })
-    listen(key.value)
+    keyListener(key.value)
     function trigger(...args: unknown[]) {
       const keyValue = key.value
       if (!keyValue) return
@@ -84,10 +94,26 @@ export function useShortcut() {
       id,
       key,
       trigger,
+      taskCount: readonly(taskCount),
+      /**
+       * `false` means that all tasks were finished, includes empty task queue and no pending task
+       */
+      taskPending: readonly(taskPending),
     }
   }
-  function pause() {}
-  function resume() {}
+  function cleanAll() {
+    Array.from(handler.map(h => h.id)).forEach(clean)
+    handler.length = 0
+  }
+  function clean(id: string) {
+    const index = handler.findIndex(h => h.id === id)
+    if (index === -1) return
+    hotkeys.unbind(handler[index].key, scope)
+    handler[index].handler.stop()
+    handler[index].queue.removeAllListeners()
+    handler[index].queue.clear()
+    handler.splice(index, 1)
+  }
   onBeforeUnmount(() => {
     cleanAll()
   })
@@ -96,7 +122,5 @@ export function useShortcut() {
     cleanAll,
     clean,
     listen,
-    resume,
-    pause,
   }
 }