+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @toolmain/shared
 
+## 1.3.1
+
+### Patch Changes
+
+- update useShortcut
+
 ## 1.3.0
 
 ### Minor Changes